# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List checklist collections
      responses:
        "200":
          description: Collections with their checklist summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Collection"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Rename a collection
      parameters:
        - $ref: "#/components/parameters/CollectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Collection not found
    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection without deleting its checklists
      parameters:
        - $ref: "#/components/parameters/CollectionId"
      responses:
        "204":
          description: Collection deleted; checklists remain available
        "404":
          description: Collection not found
  /collections/{id}/checklists:
    post:
      operationId: createCollectionChecklist
      tags: [collections]
      summary: Create a checklist inside a collection
      parameters:
        - $ref: "#/components/parameters/CollectionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistInput"
      responses:
        "201":
          description: Created checklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checklist"
        "404":
          description: Collection not found
  /collections/{id}/checklists/{checklistId}:
    put:
      operationId: addChecklistToCollection
      tags: [collections]
      summary: Add an existing checklist to a collection
      parameters:
        - $ref: "#/components/parameters/CollectionId"
        - $ref: "#/components/parameters/ChecklistIdByPath"
      responses:
        "204":
          description: Checklist added to collection
        "404":
          description: Collection or checklist not found
    delete:
      operationId: removeChecklistFromCollection
      tags: [collections]
      summary: Remove a checklist from a collection without deleting it
      parameters:
        - $ref: "#/components/parameters/CollectionId"
        - $ref: "#/components/parameters/ChecklistIdByPath"
      responses:
        "204":
          description: Checklist removed from collection
        "404":
          description: Collection or checklist not found
...
components:
  parameters:
    CollectionId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    ChecklistIdByPath:
      name: checklistId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    ChecklistStatus:
      type: string
      enum: [todo, in_progress, done, blocked]
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        checklistId:
          type: integer
        position:
          type: integer
        title:
          type: string
        note:
          type: string
        status:
          $ref: "#/components/schemas/ChecklistStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, checklistId, position, title, note, status, createdAt, updatedAt]
    Checklist:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sourceText:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalItems:
          type: integer
        completedItems:
          type: integer
        progress:
          type: integer
        isComplete:
          type: boolean
        childChecklistCount:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        relatedChecklists:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistRelation"
      required: [id, title, sourceText, createdAt, updatedAt, totalItems, completedItems, progress, isComplete, relatedChecklists, items]
    ChecklistSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        totalItems:
          type: integer
        completedItems:
          type: integer
        progress:
          type: integer
        isComplete:
          type: boolean
        childChecklistCount:
          type: integer
      required: [id, title, createdAt, updatedAt, totalItems, completedItems, progress, isComplete, childChecklistCount]
    ChecklistRelationType:
      type: string
      enum: [supports, queue, belongs_to]
    ChecklistRelation:
      type: object
      properties:
        id:
          type: integer
        parentChecklistId:
          type: integer
        childChecklistId:
          type: integer
        relationType:
          $ref: "#/components/schemas/ChecklistRelationType"
        createdAt:
          type: string
          format: date-time
        checklist:
          $ref: "#/components/schemas/ChecklistSummary"
      required: [id, parentChecklistId, childChecklistId, relationType, createdAt, checklist]
...
    Collection:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        checklists:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistSummary"
      required: [id, name, createdAt, updatedAt, checklists]
    CollectionInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
    CollectionUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
    ChecklistInput:
      type: object
      properties:
        title:
          type: string
        sourceText:
          type: string
          minLength: 1
      required: [sourceText]
...